+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+## [0.9.2] - 2024-09-22
+Minor feature release to add BetterBibtex citation and item attachment (paper)
+lookup.
+
 ### Changed
 - Refactored the database (`zensols.zotsite.db`) module by separating SQL into
   resource file and use connection pooling with the [zensols.dbutil] package.
@@ -226,7 +230,8 @@ Downstream moderate risk update release.
 - Initial version
 
 
-[Unreleased]: https://github.com/plandes/zotsite/compare/v0.9.1...HEAD
+[Unreleased]: https://github.com/plandes/zotsite/compare/v0.9.2...HEAD
+[0.9.2]: https://github.com/plandes/zotsite/compare/v0.9.1...v0.9.2
 [0.9.1]: https://github.com/plandes/zotsite/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/plandes/zotsite/compare/v0.8.2...v0.9.0
 [0.8.2]: https://github.com/plandes/zotsite/compare/v0.8.1...v0.8.2